made motion changes
fixed pid, added motion profiling for driving in a straight line, added lookahead point to pure pursuit

+++ include/helpers/rate-limiter.h
@@ -4,10 +4,10 @@
 
 class limiter{
   public: 
-    double rateLimiterOutput = 0, prevRateLimiterOutput = 0;
+    float rateLimiterOutput = 0, prevRateLimiterOutput = 0;
     
-    double rateLimiter(double val, double maxRate) {
-      double maxChange = 0.02 * maxRate;
+    float rateLimiter(float val, float maxRate) {
+      float maxChange = 0.02 * maxRate;
       rateLimiterOutput += clamp(val - prevRateLimiterOutput, -maxChange, maxChange);
       prevRateLimiterOutput = rateLimiterOutput;
       return rateLimiterOutput;
